// src/api/models/ClientesEmpresaRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientesEmpresaRequest = {
    razonSocial: string;
    numeroIdentificacion: string;
    nombreComercial?: string | null;
    nombreSucursal?: string | null;
    numeroSucursal?: string | null;
    nombreRepresentanteLegal: string;
    telefonoEmpresa: string;
    correoEmpresa: string;
    telefonoEmpresaSecundario?: string | null;
    correoEmpresaSecundario?: string | null;
};