// src/api/models/ClientesEmpresaResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientesEmpresaResponse = {
    id?: number;
    usuarioId?: number;
    razonSocial?: string | null;
    numeroIdentificacion?: string | null;
    nombreComercial?: string | null;
    nombreSucursal?: string | null;
    numeroSucursal?: string | null;
    nombreRepresentanteLegal?: string | null;
    telefonoEmpresa?: string | null;
    correoEmpresa?: string | null;
    telefonoEmpresaSecundario?: string | null;
    correoEmpresaSecundario?: string | null;
    fechaCreacion?: string;
    fechaActualizacion?: string | null;
};